WIP callnative based mid battle evolutions Something that I started to work on in my boredom. I've been wanting to do this for a long time tbh, but I'm lazy. I don't know if I'll ever finish it though.
The concept is obvious; to evolve a Pokémon in battle when they level up if they have a valid evolution path available.
I wanted to do things like this because the only public implementation uses the vanilla game's evo sequence which is a pita to go through every single time.
Updating the relevant Pokémon data and playing an animation like the one used for form changes such as Castform's feels a bit smoother to me.

To do:
-Find out why is an affected Pokémon only evolving after killing a Pokémon in the turn AFTER they leveled up.
 -Ex: Lv20 Torchic kills Lv16 Dustox, levels up, tries to learn Slash, but it's only evolving after killing a Lv16 Beautifly afterward
-Add logic to handle move learning.
-Perhaps make it so it works both, for the Player and the Opponent, idk.